lavf/rtsp: Fix a crash with the RTSP muxer.
Introduced in ca6b7ba4 / bc2a3296
The whole block that the statement was added to is only
relevant when used as a demuxer, but the other statements
there have had other if statements guarding them. Make
sure to only run this whole block if being used as a
demuxer.

Fixes ticket #5844.