webp: Fix alpha decoding
This was broken by 3aee08e4618e8ee5f32cee571b72547cb2f049ae - the webp
decoder was assuming that it could set the output pixfmt of the vp8
decoder directly, but after that change it no longer could because
ff_get_format() was used instead.  This adds an internal get_format()
callback to webp use of the vp8 decoder to override the pixfmt
appropriately.