h264: add AVOption to set x264_build default
This provides a generic way to the API user to deal with files that
either lack this SEI, or which have the SEI only in packets not passed
to the decoder (such as the common case of the SEI being in the very
firsat video packet, but decoding is started somewhere in the middle of
the file). Bugs like 7b6b6bc67cf5fe7d2fc968da94201ab6af959b71 make this
somewhat of a necessity.

This intentionally uses the version in the SEI instead, if any is found.