Revert "decode: copy the output parameters from the last bsf in the chain back to the AVCodecContext"
This reverts commit bfa8b1d6b0c065be926a9443f395cccb6538743b.

The avcodec_parameters_to_context() call was freeing and reallocating
AVCodecContext->extradata, essentially taking ownership of it, which according
to the doxy is user owned. This is an API break and has produces crashes in
some library users like Firefox.
Revert until a better solution is found to internally propagate the filtered
extradata back into the decoder context.

Signed-off-by: James Almer <jamrial@gmail.com>