avformat/mpegts: fix teletext PTS when selecting teletext streams only
After cdcb8758911a805b51e317efa4d85fef9dccca47 the teletext PTS values were set
to AV_NOPTS_VALUE if the stream of the PCR pid was discarded.

What actually matters is that if we parse the PCR of the PCR PID or not, so
let's use the cached discard value of the actual PCR PID instead of the stream
discard value, which may be different.

Also fixes ticket #7567, which was caused by the fact that teletext PTS values
were not touched if the PCR pid was discarded even before
cdcb8758911a805b51e317efa4d85fef9dccca47.

Signed-off-by: Marton Balint <cus@passwd.hu>