cbs_h265: Ensure that a predicted RPS doesn't contain too many pictures
If the RPS we are predicting from has maximum size then at least one of
the pictures in it must be discarded before adding the current one.

Also revert 5c8cae5c4b8f39070f6c688e48312042f32665c6, which added
now-redundant checks for the special case of a too-large RPS with all
pictures being in the same direction from the current one.